refactor(reply): relocate prompt budget implementation
